+++ package.json
@@ -31,7 +31,7 @@
     "zod": "^3.22.4"
   },
   "devDependencies": {
-    "daisyui": "^3.9.2",
+    "daisyui": "^4.7.2",
     "globby": "^11.0.3",
     "@svgr/cli": "^8.1.0"
   }
