+++ src/components/atoms/Table/index.tsx
@@ -1,9 +1,10 @@
-import { ChangeEvent, ReactNode, useState } from "react"
+'use client'
+import { ReactNode } from "react"
 
 import clsx from "clsx"
 
-import { ChevronLeft, ChevronRight } from "../../icons"
-import styles from "./Table.module.scss"
+import Pagination from "./Pagination"
+import styles from "./styles.module.css"
 
 export interface Pagination {
   total: number
@@ -44,29 +45,26 @@ const Table = <T extends {}>({
   onChangeSort = () => void 0,
   onChangeItemsPerPage = () => void 0,
 }: TableProps<T>) => {
-  const [currentPage, setCurrentPage] = useState(0)
-  const [rowsPerPage, setRowsPerPage] = useState(
-    rowsPerPageOptions ? rowsPerPageOptions[0] : 10
-  )
+  
 
   const initialSortColumn = columns.find(
     (col) => col.initialSortDirection != null
   )
-  const [sortConfig, setSortConfig] = useState<{
-    dataIndex?: string | null
-    direction?: "asc" | "desc" | null
-  }>({
-    dataIndex: initialSortColumn ? initialSortColumn.dataIndex : null,
-    direction: initialSortColumn
-      ? initialSortColumn.initialSortDirection
-      : null,
-  })
+  // const [sortConfig, setSortConfig] = useState<{
+  //   dataIndex?: string | null
+  //   direction?: "asc" | "desc" | null
+  // }>({
+  //   dataIndex: initialSortColumn ? initialSortColumn.dataIndex : null,
+  //   direction: initialSortColumn
+  //     ? initialSortColumn.initialSortDirection
+  //     : null,
+  // })
 
   const activeColumns = columns.filter((column) =>
     data.some((row) => row[column.dataIndex])
   )
 
-  const sortedRowsPerPageOptions = [...rowsPerPageOptions].sort((a, b) => a - b)
+  
 
   const tableRows = data?.map((row, rowIndex) => (
     <tr
@@ -92,92 +90,17 @@ const Table = <T extends {}>({
     </tr>
   ))
 
-  // const handleChangePage = (page: number) => {
-  //     setCurrentPage(page)
+  // const handleSortClick = (dataIndex: string) => {
+  //   setSortConfig((prevConfig) => {
+  //     let newDirection = "asc"
+  //     if (prevConfig.dataIndex === dataIndex) {
+  //       newDirection = prevConfig.direction === "asc" ? "desc" : "asc"
+  //     }
+  //     onChangeSort(dataIndex, newDirection)
+
+  //     return { dataIndex, direction: newDirection }
+  //   })
   // }
-
-  const numPages = Math.ceil((pagination!.total || rowsPerPage) / rowsPerPage)
-
-  const handleChangeRowsPerPage = (event: ChangeEvent<HTMLSelectElement>) => {
-    const newRowsPerPage = parseInt(event.target.value, 10)
-    setRowsPerPage(newRowsPerPage)
-    handleCurrentPage(0)
-    onChangeItemsPerPage(newRowsPerPage)
-  }
-
-  const handleChangePageInput = (event: ChangeEvent<HTMLInputElement>) => {
-    const pageNumber = parseInt(event.target.value, 10) - 1
-    if (pageNumber >= 0 && pageNumber < numPages) {
-      handleCurrentPage(pageNumber)
-    }
-  }
-
-  const handleSortClick = (dataIndex: string) => {
-    setSortConfig((prevConfig) => {
-      let newDirection = "asc"
-      if (prevConfig.dataIndex === dataIndex) {
-        newDirection = prevConfig.direction === "asc" ? "desc" : "asc"
-      }
-      onChangeSort(dataIndex, newDirection)
-
-      return { dataIndex, direction: newDirection }
-    })
-  }
-
-  const handleCurrentPage = (page: number) => {
-    setCurrentPage(page)
-    onChangePage(page)
-  }
-
-  const paginationControls = pagination && (
-    <div className={styles.pagination}>
-      <div className={styles.pageNumbers}>
-        {/* {Array.from({length: numPages}, (_, i) => (
-                    <button
-                        key={i}
-                        className={`${styles.pageNumber} ${i === currentPage ? 'active' : ''}`}
-                        onClick={() => handleChangePage(i)}
-                    >
-                        {i + 1}
-                    </button>
-                ))} */}
-        <button
-          onClick={() => handleCurrentPage(currentPage - 1)}
-          disabled={currentPage === 0}
-        >
-          {<ChevronLeft />}
-        </button>
-        <input
-          type="number"
-          min="1"
-          max={numPages}
-          value={currentPage + 1}
-          onChange={handleChangePageInput}
-          className={styles.pageInput}
-          title="Page Number"
-        />
-        <span>of {numPages}</span>
-        <button
-          onClick={() => handleCurrentPage(currentPage + 1)}
-          disabled={currentPage === numPages - 1}
-        >
-          {<ChevronRight />}
-        </button>
-      </div>
-      {rowsPerPageOptions && rowsPerPageOptions.length > 0 && (
-        <div className={styles.tableRecordsSelector}>
-          <select value={rowsPerPage} onChange={handleChangeRowsPerPage}>
-            {sortedRowsPerPageOptions.map((option) => (
-              <option key={option} value={option}>
-                {option}
-              </option>
-            ))}
-          </select>
-          <span>&nbsp;items</span>
-        </div>
-      )}
-    </div>
-  )
 
   return (
     <div className={styles.tableContainer}>
@@ -210,7 +133,7 @@ const Table = <T extends {}>({
                     styles.headerCellContent,
                     col.sortable && styles.sortable
                   )}
-                  onClick={() => col.sortable && handleSortClick(col.dataIndex)}
+                  // onClick={() => col.sortable && handleSortClick(col.dataIndex)}
                 >
                   {col.renderHeader ? col.renderHeader(col.title) : col.title}
                   {col.sortable && (
@@ -233,7 +156,7 @@ const Table = <T extends {}>({
         </thead>
         <tbody className={styles.tableBody}>{tableRows}</tbody>
       </table>
-      {paginationControls}
+      <Pagination/>
     </div>
   )
 }
